SIG-935 moved most of the API code out of signals.apps.signals
