+++ api/app/signals/apps/api/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'signals.apps.api.config.ApiConfig'
