+++ api/app/signals/apps/api/config.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ApiConfig(AppConfig):
+    name = 'signals.apps.api'
+    verbose_name = 'REST API App'
