+++ api/app/signals/apps/api/fixtures/category_auth_post.json
