+++ api/app/signals/apps/api/fixtures/location_auth_post.json
