+++ api/app/signals/apps/api/fixtures/status_auth_post.json
