+++ api/app/signals/apps/api/pdf/mixins.py
