+++ api/app/signals/apps/api/templates/api/print_signal.html
