+++ api/app/signals/apps/api/templates/api/swagger/openapi.yaml
