+++ api/app/signals/apps/api/templates/rest_framework/api.html
