+++ api/app/signals/urls.py
@@ -8,7 +8,7 @@ urlpatterns = [
     path('status/', include('signals.apps.health.urls')),
 
     # The Signals application is routed behind the HAproxy with `/signals/` as path.
-    path('signals/', include('signals.apps.signals.urls')),
+    path('signals/', include('signals.apps.api.urls')),
     path('signals/user/', include('signals.apps.users.urls')),
     path('signals/admin/', admin.site.urls),
