+++ api/app/tests/apps/signals/test_filters.py
@@ -8,9 +8,9 @@ from rest_framework.mixins import ListModelMixin
 from rest_framework.pagination import LimitOffsetPagination
 from rest_framework.viewsets import GenericViewSet
 
-from signals.apps.signals.api_generics.filters import FieldMappingOrderingFilter
+from signals.apps.api.generics.filters import FieldMappingOrderingFilter
+from signals.apps.api.v0.serializers import SignalAuthHALSerializer
 from signals.apps.signals.models import Category, MainCategory, Priority, Signal
-from signals.apps.signals.v0.serializers import SignalAuthHALSerializer
 from tests.apps.signals.factories import SignalFactory
 from tests.test import SignalsBaseApiTestCase
